+++ internal/agent/tools/browser.go
@@ -1,4 +1,4 @@
-package browser
+package tools
 
 import (
 	"bytes"
@@ -14,17 +14,12 @@ import (
 
 	"github.com/memohai/memoh/internal/browsercontexts"
 	"github.com/memohai/memoh/internal/config"
-	mcpgw "github.com/memohai/memoh/internal/mcp"
 	"github.com/memohai/memoh/internal/settings"
 	"github.com/memohai/memoh/internal/workspace/bridge"
+	sdk "github.com/memohai/twilight-ai/sdk"
 )
 
-const (
-	toolBrowserAction  = "browser_action"
-	toolBrowserObserve = "browser_observe"
-)
-
-type Executor struct {
+type BrowserProvider struct {
 	logger          *slog.Logger
 	settings        *settings.Service
 	browserContexts *browsercontexts.Service
@@ -33,12 +28,12 @@ type Executor struct {
 	httpClient      *http.Client
 }
 
-func NewExecutor(log *slog.Logger, settingsSvc *settings.Service, browserSvc *browsercontexts.Service, containers bridge.Provider, gatewayCfg config.BrowserGatewayConfig) *Executor {
+func NewBrowserProvider(log *slog.Logger, settingsSvc *settings.Service, browserSvc *browsercontexts.Service, containers bridge.Provider, gatewayCfg config.BrowserGatewayConfig) *BrowserProvider {
 	if log == nil {
 		log = slog.Default()
 	}
-	return &Executor{
-		logger:          log.With(slog.String("provider", "browser_tool")),
+	return &BrowserProvider{
+		logger:          log.With(slog.String("tool", "browser")),
 		settings:        settingsSvc,
 		browserContexts: browserSvc,
 		containers:      containers,
@@ -47,26 +42,27 @@ func NewExecutor(log *slog.Logger, settingsSvc *settings.Service, browserSvc *br
 	}
 }
 
-func (e *Executor) ListTools(ctx context.Context, session mcpgw.ToolSessionContext) ([]mcpgw.ToolDescriptor, error) {
-	if e.settings == nil || e.browserContexts == nil {
-		return []mcpgw.ToolDescriptor{}, nil
+func (p *BrowserProvider) Tools(ctx context.Context, session SessionContext) ([]sdk.Tool, error) {
+	if p.settings == nil || p.browserContexts == nil {
+		return nil, nil
 	}
 	botID := strings.TrimSpace(session.BotID)
 	if botID == "" {
-		return []mcpgw.ToolDescriptor{}, nil
+		return nil, nil
 	}
-	botSettings, err := e.settings.GetBot(ctx, botID)
+	botSettings, err := p.settings.GetBot(ctx, botID)
 	if err != nil {
-		return []mcpgw.ToolDescriptor{}, nil
+		return nil, nil
 	}
 	if strings.TrimSpace(botSettings.BrowserContextID) == "" {
-		return []mcpgw.ToolDescriptor{}, nil
+		return nil, nil
 	}
-	return []mcpgw.ToolDescriptor{
+	sess := session
+	return []sdk.Tool{
 		{
-			Name:        toolBrowserAction,
+			Name:        "browser_action",
 			Description: "Execute a browser action: navigate, click, double-click, focus, type, fill, press key, keyboard input, hover, select option, check/uncheck, scroll, drag-and-drop, upload files, go back/forward, reload, wait, or manage tabs (new/select/close).",
-			InputSchema: map[string]any{
+			Parameters: map[string]any{
 				"type": "object",
 				"properties": map[string]any{
 					"action":          map[string]any{"type": "string", "enum": []string{"navigate", "click", "dblclick", "focus", "type", "fill", "press", "keyboard_type", "keyboard_inserttext", "keydown", "keyup", "hover", "select", "check", "uncheck", "scroll", "scrollintoview", "drag", "upload", "wait", "go_back", "go_forward", "reload", "tab_new", "tab_select", "tab_close"}, "description": "The browser action to perform"},
@@ -84,11 +80,14 @@ func (e *Executor) ListTools(ctx context.Context, session mcpgw.ToolSessionConte
 				},
 				"required": []string{"action"},
 			},
+			Execute: func(ctx *sdk.ToolExecContext, input any) (any, error) {
+				return p.execAction(ctx, sess, inputAsMap(input))
+			},
 		},
 		{
-			Name:        toolBrowserObserve,
+			Name:        "browser_observe",
 			Description: "Observe the current browser page: take screenshot (optionally annotated with numbered element labels or full-page), get accessibility tree snapshot, get text content, get HTML, evaluate JavaScript, get current URL, get page title, export PDF, or list open tabs.",
-			InputSchema: map[string]any{
+			Parameters: map[string]any{
 				"type": "object",
 				"properties": map[string]any{
 					"observe":   map[string]any{"type": "string", "enum": []string{"screenshot", "screenshot_annotate", "snapshot", "get_content", "get_html", "evaluate", "get_url", "get_title", "pdf", "tab_list"}, "description": "What to observe from the page"},
@@ -98,60 +97,104 @@ func (e *Executor) ListTools(ctx context.Context, session mcpgw.ToolSessionConte
 				},
 				"required": []string{"observe"},
 			},
+			Execute: func(ctx *sdk.ToolExecContext, input any) (any, error) {
+				return p.execObserve(ctx, sess, inputAsMap(input))
+			},
 		},
 	}, nil
 }
 
-func (e *Executor) CallTool(ctx context.Context, session mcpgw.ToolSessionContext, toolName string, arguments map[string]any) (map[string]any, error) {
-	if e.settings == nil || e.browserContexts == nil {
-		return mcpgw.BuildToolErrorResult("browser tools are not available"), nil
-	}
-	botID := strings.TrimSpace(session.BotID)
-	if botID == "" {
-		return mcpgw.BuildToolErrorResult("bot_id is required"), nil
-	}
-
-	botSettings, err := e.settings.GetBot(ctx, botID)
+func (p *BrowserProvider) resolveContext(ctx context.Context, botID string) (string, browsercontexts.BrowserContext, error) {
+	botSettings, err := p.settings.GetBot(ctx, botID)
 	if err != nil {
-		return mcpgw.BuildToolErrorResult(err.Error()), nil
+		return "", browsercontexts.BrowserContext{}, err
 	}
 	browserCtxID := strings.TrimSpace(botSettings.BrowserContextID)
 	if browserCtxID == "" {
-		return mcpgw.BuildToolErrorResult("browser context not configured for this bot"), nil
+		return "", browsercontexts.BrowserContext{}, fmt.Errorf("browser context not configured for this bot")
 	}
-
-	bcConfig, err := e.browserContexts.GetByID(ctx, browserCtxID)
+	bcConfig, err := p.browserContexts.GetByID(ctx, browserCtxID)
 	if err != nil {
-		return mcpgw.BuildToolErrorResult("failed to load browser context config: " + err.Error()), nil
+		return "", browsercontexts.BrowserContext{}, fmt.Errorf("failed to load browser context config: %s", err.Error())
 	}
-
-	if err := e.ensureContext(ctx, browserCtxID, bcConfig); err != nil {
-		return mcpgw.BuildToolErrorResult("failed to ensure browser context: " + err.Error()), nil
+	if err := p.ensureContext(ctx, browserCtxID, bcConfig); err != nil {
+		return "", browsercontexts.BrowserContext{}, fmt.Errorf("failed to ensure browser context: %s", err.Error())
 	}
-
-	switch toolName {
-	case toolBrowserAction:
-		return e.callAction(ctx, botID, browserCtxID, arguments)
-	case toolBrowserObserve:
-		return e.callObserve(ctx, botID, browserCtxID, arguments)
-	default:
-		return nil, mcpgw.ErrToolNotFound
-	}
+	return browserCtxID, bcConfig, nil
 }
 
-func (e *Executor) ensureContext(ctx context.Context, contextID string, bc browsercontexts.BrowserContext) error {
-	existsURL := fmt.Sprintf("%s/context/%s/exists", e.gatewayBaseURL, contextID)
+func (p *BrowserProvider) execAction(ctx context.Context, session SessionContext, args map[string]any) (any, error) {
+	botID := strings.TrimSpace(session.BotID)
+	if botID == "" {
+		return nil, fmt.Errorf("bot_id is required")
+	}
+	contextID, _, err := p.resolveContext(ctx, botID)
+	if err != nil {
+		return nil, err
+	}
+	action := StringArg(args, "action")
+	if action == "" {
+		return nil, fmt.Errorf("action is required")
+	}
+	payload := map[string]any{"action": action}
+	for _, key := range []string{"url", "selector", "text", "key", "value", "target_selector", "direction"} {
+		if v := StringArg(args, key); v != "" {
+			payload[key] = v
+		}
+	}
+	if v, ok, _ := IntArg(args, "timeout"); ok {
+		payload["timeout"] = v
+	}
+	if v, ok, _ := IntArg(args, "amount"); ok {
+		payload["amount"] = v
+	}
+	if v, ok, _ := IntArg(args, "tab_index"); ok {
+		payload["tab_index"] = v
+	}
+	if files, ok := args["files"].([]any); ok && len(files) > 0 {
+		payload["files"] = files
+	}
+	return p.doGatewayAction(ctx, botID, contextID, payload)
+}
+
+func (p *BrowserProvider) execObserve(ctx context.Context, session SessionContext, args map[string]any) (any, error) {
+	botID := strings.TrimSpace(session.BotID)
+	if botID == "" {
+		return nil, fmt.Errorf("bot_id is required")
+	}
+	contextID, _, err := p.resolveContext(ctx, botID)
+	if err != nil {
+		return nil, err
+	}
+	observe := StringArg(args, "observe")
+	if observe == "" {
+		return nil, fmt.Errorf("observe is required")
+	}
+	payload := map[string]any{"action": observe}
+	if v := StringArg(args, "selector"); v != "" {
+		payload["selector"] = v
+	}
+	if v := StringArg(args, "script"); v != "" {
+		payload["script"] = v
+	}
+	if v, ok := args["full_page"].(bool); ok {
+		payload["full_page"] = v
+	}
+	return p.doGatewayAction(ctx, botID, contextID, payload)
+}
+
+func (p *BrowserProvider) ensureContext(ctx context.Context, contextID string, bc browsercontexts.BrowserContext) error {
+	existsURL := fmt.Sprintf("%s/context/%s/exists", p.gatewayBaseURL, contextID)
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, existsURL, nil)
 	if err != nil {
 		return err
 	}
-	resp, err := e.httpClient.Do(req) //nolint:gosec // URL from internal gateway config
+	resp, err := p.httpClient.Do(req) //nolint:gosec
 	if err != nil {
 		return fmt.Errorf("browser gateway unreachable: %w", err)
 	}
 	defer func() { _ = resp.Body.Close() }()
 	body, _ := io.ReadAll(resp.Body)
-
 	var existsResp struct {
 		Exists bool `json:"exists"`
 	}
@@ -161,19 +204,14 @@ func (e *Executor) ensureContext(ctx context.Context, contextID string, bc brows
 	if existsResp.Exists {
 		return nil
 	}
-
-	createPayload, _ := json.Marshal(map[string]any{
-		"id":     contextID,
-		"name":   bc.Name,
-		"config": bc.Config,
-	})
-	createURL := fmt.Sprintf("%s/context", e.gatewayBaseURL)
+	createPayload, _ := json.Marshal(map[string]any{"id": contextID, "name": bc.Name, "config": bc.Config})
+	createURL := fmt.Sprintf("%s/context", p.gatewayBaseURL)
 	createReq, err := http.NewRequestWithContext(ctx, http.MethodPost, createURL, bytes.NewReader(createPayload))
 	if err != nil {
 		return err
 	}
 	createReq.Header.Set("Content-Type", "application/json")
-	createResp, err := e.httpClient.Do(createReq) //nolint:gosec // URL from internal gateway config
+	createResp, err := p.httpClient.Do(createReq) //nolint:gosec
 	if err != nil {
 		return fmt.Errorf("failed to create browser context: %w", err)
 	}
@@ -185,96 +223,47 @@ func (e *Executor) ensureContext(ctx context.Context, contextID string, bc brows
 	return nil
 }
 
-func (e *Executor) callAction(ctx context.Context, botID, contextID string, arguments map[string]any) (map[string]any, error) {
-	action := mcpgw.StringArg(arguments, "action")
-	if action == "" {
-		return mcpgw.BuildToolErrorResult("action is required"), nil
-	}
-	payload := map[string]any{"action": action}
-	for _, key := range []string{"url", "selector", "text", "key", "value", "target_selector", "direction"} {
-		if v := mcpgw.StringArg(arguments, key); v != "" {
-			payload[key] = v
-		}
-	}
-	if v, ok, _ := mcpgw.IntArg(arguments, "timeout"); ok {
-		payload["timeout"] = v
-	}
-	if v, ok, _ := mcpgw.IntArg(arguments, "amount"); ok {
-		payload["amount"] = v
-	}
-	if v, ok, _ := mcpgw.IntArg(arguments, "tab_index"); ok {
-		payload["tab_index"] = v
-	}
-	if files, ok := arguments["files"].([]any); ok && len(files) > 0 {
-		payload["files"] = files
-	}
-	return e.doGatewayAction(ctx, botID, contextID, payload)
-}
-
-func (e *Executor) callObserve(ctx context.Context, botID, contextID string, arguments map[string]any) (map[string]any, error) {
-	observe := mcpgw.StringArg(arguments, "observe")
-	if observe == "" {
-		return mcpgw.BuildToolErrorResult("observe is required"), nil
-	}
-	payload := map[string]any{"action": observe}
-	if v := mcpgw.StringArg(arguments, "selector"); v != "" {
-		payload["selector"] = v
-	}
-	if v := mcpgw.StringArg(arguments, "script"); v != "" {
-		payload["script"] = v
-	}
-	if v, ok := arguments["full_page"].(bool); ok {
-		payload["full_page"] = v
-	}
-	return e.doGatewayAction(ctx, botID, contextID, payload)
-}
-
-func (e *Executor) doGatewayAction(ctx context.Context, botID, contextID string, payload map[string]any) (map[string]any, error) {
+func (p *BrowserProvider) doGatewayAction(ctx context.Context, botID, contextID string, payload map[string]any) (any, error) {
 	body, _ := json.Marshal(payload)
-	actionURL := fmt.Sprintf("%s/context/%s/action", e.gatewayBaseURL, contextID)
+	actionURL := fmt.Sprintf("%s/context/%s/action", p.gatewayBaseURL, contextID)
 	req, err := http.NewRequestWithContext(ctx, http.MethodPost, actionURL, bytes.NewReader(body))
 	if err != nil {
-		return mcpgw.BuildToolErrorResult(err.Error()), nil
+		return nil, err
 	}
 	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := e.httpClient.Do(req) //nolint:gosec // URL from internal gateway config
+	resp, err := p.httpClient.Do(req) //nolint:gosec
 	if err != nil {
-		return mcpgw.BuildToolErrorResult("browser gateway request failed: " + err.Error()), nil
+		return nil, fmt.Errorf("browser gateway request failed: %s", err.Error())
 	}
 	defer func() { _ = resp.Body.Close() }()
 	respBody, _ := io.ReadAll(resp.Body)
-
 	var gwResp struct {
 		Success bool           `json:"success"`
 		Data    map[string]any `json:"data"`
 		Error   string         `json:"error"`
 	}
 	if err := json.Unmarshal(respBody, &gwResp); err != nil {
-		return mcpgw.BuildToolErrorResult("invalid gateway response"), nil
+		return nil, fmt.Errorf("invalid gateway response")
 	}
 	if !gwResp.Success {
 		errMsg := gwResp.Error
 		if errMsg == "" {
 			errMsg = "browser action failed"
 		}
-		return mcpgw.BuildToolErrorResult(errMsg), nil
+		return nil, fmt.Errorf("%s", errMsg)
 	}
-
 	if b64, ok := gwResp.Data["screenshot"].(string); ok && b64 != "" {
-		return e.buildScreenshotResult(ctx, botID, b64), nil
+		return p.buildScreenshotResult(ctx, botID, b64), nil
 	}
-	return mcpgw.BuildToolSuccessResult(gwResp.Data), nil
+	return gwResp.Data, nil
 }
 
-const screenshotContainerDir = "/data/browser-screenshots"
+const browserScreenshotDir = "/data/browser-screenshots"
 
-func (e *Executor) buildScreenshotResult(ctx context.Context, botID, base64Data string) map[string]any {
+func (p *BrowserProvider) buildScreenshotResult(ctx context.Context, botID, base64Data string) any {
 	mimeType := "image/png"
-
 	imgBytes, err := base64.StdEncoding.DecodeString(base64Data)
 	if err != nil {
-		e.logger.Warn("failed to decode screenshot base64", slog.Any("error", err))
 		return map[string]any{
 			"content": []map[string]any{
 				{"type": "text", "text": "Screenshot captured (failed to decode for saving)"},
@@ -282,12 +271,9 @@ func (e *Executor) buildScreenshotResult(ctx context.Context, botID, base64Data 
 			},
 		}
 	}
-
-	containerPath := fmt.Sprintf("%s/%d.png", screenshotContainerDir, time.Now().UnixMilli())
-
-	client, clientErr := e.containers.MCPClient(ctx, botID)
+	containerPath := fmt.Sprintf("%s/%d.png", browserScreenshotDir, time.Now().UnixMilli())
+	client, clientErr := p.containers.MCPClient(ctx, botID)
 	if clientErr != nil {
-		e.logger.Warn("container not reachable for screenshot save", slog.String("bot_id", botID), slog.Any("error", clientErr))
 		return map[string]any{
 			"content": []map[string]any{
 				{"type": "text", "text": "Screenshot captured (container not reachable, not saved to disk)"},
@@ -295,12 +281,9 @@ func (e *Executor) buildScreenshotResult(ctx context.Context, botID, base64Data 
 			},
 		}
 	}
-
-	mkdirCmd := fmt.Sprintf("mkdir -p %s", screenshotContainerDir)
+	mkdirCmd := fmt.Sprintf("mkdir -p %s", browserScreenshotDir)
 	_, _ = client.Exec(ctx, mkdirCmd, "/", 5)
-
 	if writeErr := client.WriteFile(ctx, containerPath, imgBytes); writeErr != nil {
-		e.logger.Warn("failed to write screenshot to container", slog.String("bot_id", botID), slog.Any("error", writeErr))
 		return map[string]any{
 			"content": []map[string]any{
 				{"type": "text", "text": fmt.Sprintf("Screenshot captured (failed to save: %s)", writeErr.Error())},
@@ -308,7 +291,6 @@ func (e *Executor) buildScreenshotResult(ctx context.Context, botID, base64Data 
 			},
 		}
 	}
-
 	return map[string]any{
 		"content": []map[string]any{
 			{"type": "text", "text": fmt.Sprintf("Screenshot saved to %s", containerPath)},
