+++ internal/mcp/tool_gateway_service.go
@@ -18,26 +18,21 @@ type cachedToolRegistry struct {
 	registry  *ToolRegistry
 }
 
-// ToolGatewayService federates tools from executors and sources.
+// ToolGatewayService federates tools from external MCP sources (federation).
+// Built-in tools are no longer managed here — they are loaded directly
+// via agent ToolProviders.
 type ToolGatewayService struct {
-	logger    *slog.Logger
-	executors []ToolExecutor
-	sources   []ToolSource
-	cacheTTL  time.Duration
+	logger   *slog.Logger
+	sources  []ToolSource
+	cacheTTL time.Duration
 
 	mu    sync.Mutex
 	cache map[string]cachedToolRegistry
 }
 
-func NewToolGatewayService(log *slog.Logger, executors []ToolExecutor, sources []ToolSource) *ToolGatewayService {
+func NewToolGatewayService(log *slog.Logger, sources []ToolSource) *ToolGatewayService {
 	if log == nil {
 		log = slog.Default()
-	}
-	filteredExecutors := make([]ToolExecutor, 0, len(executors))
-	for _, executor := range executors {
-		if executor != nil {
-			filteredExecutors = append(filteredExecutors, executor)
-		}
 	}
 	filteredSources := make([]ToolSource, 0, len(sources))
 	for _, source := range sources {
@@ -46,11 +41,10 @@ func NewToolGatewayService(log *slog.Logger, executors []ToolExecutor, sources [
 		}
 	}
 	return &ToolGatewayService{
-		logger:    log.With(slog.String("service", "tool_gateway")),
-		executors: filteredExecutors,
-		sources:   filteredSources,
-		cacheTTL:  defaultToolRegistryCacheTTL,
-		cache:     map[string]cachedToolRegistry{},
+		logger:   log.With(slog.String("service", "tool_gateway")),
+		sources:  filteredSources,
+		cacheTTL: defaultToolRegistryCacheTTL,
+		cache:    map[string]cachedToolRegistry{},
 	}
 }
 
@@ -87,14 +81,13 @@ func (s *ToolGatewayService) CallTool(ctx context.Context, session ToolSessionCo
 	if err != nil {
 		return nil, err
 	}
-	executor, _, ok := registry.Lookup(toolName)
+	source, _, ok := registry.Lookup(toolName)
 	if !ok {
-		// Refresh once for dynamic executors/sources.
 		registry, err = s.getRegistry(ctx, session, true)
 		if err != nil {
 			return nil, err
 		}
-		executor, _, ok = registry.Lookup(toolName)
+		source, _, ok = registry.Lookup(toolName)
 		if !ok {
 			return BuildToolErrorResult("tool not found: " + toolName), nil
 		}
@@ -104,7 +97,7 @@ func (s *ToolGatewayService) CallTool(ctx context.Context, session ToolSessionCo
 	if arguments == nil {
 		arguments = map[string]any{}
 	}
-	result, err := executor.CallTool(ctx, session, toolName, arguments)
+	result, err := source.CallTool(ctx, session, toolName, arguments)
 	if err != nil {
 		if errors.Is(err, ErrToolNotFound) {
 			return BuildToolErrorResult("tool not found: " + toolName), nil
@@ -133,18 +126,6 @@ func (s *ToolGatewayService) getRegistry(ctx context.Context, session ToolSessio
 	}
 
 	registry := NewToolRegistry()
-	for _, executor := range s.executors {
-		tools, err := executor.ListTools(ctx, session)
-		if err != nil {
-			s.logger.Warn("list tools from executor failed", slog.Any("error", err))
-			continue
-		}
-		for _, tool := range tools {
-			if err := registry.Register(executor, tool); err != nil {
-				s.logger.Warn("skip duplicated/invalid tool", slog.String("tool", tool.Name), slog.Any("error", err))
-			}
-		}
-	}
 	for _, source := range s.sources {
 		tools, err := source.ListTools(ctx, session)
 		if err != nil {
